replaced sin and cos with table lookups, drawCircle no longer indexes off a float

+++ excalibur/exMap.h
@@ -177,7 +177,7 @@ protected:
   exTimeType    _last_fps;
   QTimer idleTimer;
   
-  void drawCircle(int radius, uint8_t segments);
+  void drawCircle(int radius, int segments);
   void drawEXMob(exMob *m);
   void drawMobName(exMob *m);
   void drawAggroCircle(GLfloat R, GLfloat G, GLfloat B, GLfloat distfade_pct);
